// Microsoft.MyWorkspace.UI/src/types/enums/DashboardCards.ts
export enum DashboardCards {
  None = 0,
  WorkspaceAgeStalenessCard = 1,
  WorkspaceStatusOverviewCard = 2,
  JitStatusOverviewCard = 3,
  DocumentationCard = 4,
  WorkspaceShutdownScheduleCard = 5,
  WorkspaceCountOverviewCard = 6,
  TourCard = 7,
  WorkspaceDeletionCard = 8,
  FeatureAnnouncementCard = 9,
  ComplianceUpdateCard = 10,
}